Rename artifacts, packages, binaries and prepare CI
