+++ .azure/templates/setup_java.yaml
@@ -1,0 +1,40 @@
+# Step to setup JAVA on the agent
+# We use openjdk-X, where X is Java version (8 or 11). Images are based on Java 8
+parameters:
+  - name: JDK_PATH
+    default: '/usr/lib/jvm/java-8-openjdk-amd64'
+  - name: JDK_VERSION
+    default: '1.8'
+steps:
+
+- bash: |
+    sudo apt-get update
+  displayName: 'Update package list'
+
+- bash: |
+    sudo apt-get install openjdk-8-jdk
+  displayName: 'Install openjdk8'
+  condition: eq(variables['JDK_VERSION'], '1.8')
+
+- bash: |
+    sudo apt-get install openjdk-11-jdk
+  displayName: 'Install openjdk11'
+  condition: eq(variables['JDK_VERSION'], '11')
+
+- bash: |
+    echo "##vso[task.setvariable variable=JAVA_VERSION_BUILD]1.8"
+    echo "##vso[task.setvariable variable=JAVA_VERSION]1.8.0"
+  displayName: 'Setup JAVA_VERSION=1.8'
+  condition: eq(variables['JDK_VERSION'], '1.8')
+
+- bash: |
+    echo "##vso[task.setvariable variable=JAVA_VERSION_BUILD]11"
+    echo "##vso[task.setvariable variable=JAVA_VERSION]11"
+  displayName: 'Setup JAVA_VERSION=11'
+  condition: eq(variables['JDK_VERSION'], '11')
+
+- bash: |
+    echo "##vso[task.setvariable variable=JAVA_HOME]$(JDK_PATH)"
+    echo "##vso[task.setvariable variable=JAVA_HOME__X64]$(JDK_PATH)"
+    echo "##vso[task.setvariable variable=PATH]$(jdk_path)/bin:$(PATH)"
+  displayName: 'Setup JAVA_HOME'
