+++ .azure/templates/setup_minikube.yaml
@@ -1,0 +1,15 @@
+# Steps needed for local Minikube instance
+steps:
+- task: DockerInstaller@0
+  displayName: Docker Installer
+  inputs:
+    dockerVersion: 19.03.9
+    releaseType: stable
+- bash: ".azure/scripts/setup-kubernetes.sh"
+  displayName: "Setup Minikube cluster"
+  env:
+    KUBE_CLUSTER: 'minikube'
+    KUBE_VERSION: 'latest'
+    KUBECTL_VERSION: 'latest'
+    MINIKUBE_VERSION: 'latest'
+    NSENTER_VERSION: '2.32'
