--- src/main/docker/Dockerfile.native-distroless
@@ -1,23 +1,0 @@
-####
-# This Dockerfile is used in order to build a distroless container that runs the Quarkus application in native (no JVM) mode
-#
-# Before building the container image run:
-#
-# ./mvnw package -Pnative
-#
-# Then, build the image with:
-#
-# docker build -f src/main/docker/Dockerfile.native-distroless -t quarkus/code-with-quarkus .
-#
-# Then run the container using:
-#
-# docker run -i --rm -p 8080:8080 quarkus/code-with-quarkus
-#
-###
-FROM quay.io/quarkus/quarkus-distroless-image:1.0
-COPY target/*-runner /application
-
-EXPOSE 8080
-USER nonroot
-
-CMD ["./application", "-Dquarkus.http.host=0.0.0.0"]
