+++ src/main/java/io/strimzi/ValidatingWebhook.java
@@ -1,4 +1,4 @@
-package cz.scholz;
+package io.strimzi;
 
 import io.fabric8.kubernetes.api.model.Pod;
 import io.fabric8.kubernetes.api.model.admission.v1.AdmissionRequest;
