+++ src/test/java/io/strimzi/DrainCleanerTest.java
@@ -1,4 +1,4 @@
-package cz.scholz;
+package io.strimzi;
 
 import io.fabric8.kubernetes.api.model.Pod;
 import io.fabric8.kubernetes.api.model.PodBuilder;
